# backend/src/docs/api-docs.yaml
openapi: 3.0.0
info:
  title: HMS Backend Code Wizards
  version: 1.0.0
  description:  Description of all the paths of the backend as well as needed information and status codes
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/course/{courseCode}:
    get:
      summary: Find details for a certain course
      operationId: getCourse
      tags:
        - Course
      parameters:
        - name: courseCode
          in: path
          required: true
          description: The course code to find details about
          schema:
            type: string
      responses:
        '200':
          description: Courses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Course'
                type: object
        '404':
          description: Course not found
        '500':
          description: Internal server error    
...
  /api/assignment/course/{courseCode}:
    get:
      summary: Get all assignments for a course
      operationId: getAllAssignments
      tags:
        - Assignment
      responses:
        '200':
          description: A list of assignments retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Assignment'
                type: array
        '404':
          description: Assignments not found
        '500':
          description: Internal server error
  /api/assignment/{assignCode}:
    get:
      summary: Get details for an assignment
      operationId: getAssignments
      tags:
        - Assignment
      responses:
        '200':
          description: Details for assignment successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Assignment'
                type: object
        '404':
          description: Assignments not found
        '500':
          description: Internal server error
  /api/assignment/create:
    post:
      summary: Create a new assignment
      operationId: createAssignment
      security:
        - bearerAuth: []
      tags:
        - Assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                dueDate:
                  type: string
                  format: date
                courseCode:
                  type: string
                  example: CS101
      responses:
        '200':
          description: Assignment created successfully
        '400': 
          description: Bad request
        '401':
          description: Unauthorized - Missing or invalid token
        '404':
          description: Course not found
        '500':
          description: Internal server error
  /api/assignment/update/{assignCode}:
    put:
      summary: Update an assignment
      operationId: updateAssignment
      security:
        - bearerAuth: []
      tags:
        - Assignment
      parameters:
        - name: assignCode
          in: path
          required: true
          description: The assignment code to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                dueDate:
                  type: string
                  format: date
      responses:
        '200':
          description: Assignment updated successfully
        '400': 
          description: Bad request
        '401':
          description: Unauthorized - Missing or invalid token
        '404':
          description: Assignment not found
        '500':
          description: Internal server error
  /api/assignment/delete/{assignCode}:
    delete:
      summary: Delete an assignment
      operationId: deleteAssignment
      security:
        - bearerAuth: []
      tags:
        - Assignment
      parameters:
        - name: assignCode
          in: path
          required: true
          description: The assignment code to delete
          schema:
            type: string
      responses:
        '200':
          description: Assignment deleted successfully
        '400': 
          description: Bad request
        '401':
          description: Unauthorized - Missing or invalid token
        '404':
          description: Assignment not found
        '500':
          description: Internal server error
...
  /api/submission/assignment/{assignCode}:
    get:
      summary: Get submissions for an assignment
      operationId: getSubmissionsByAssignment
      tags:
        - Submission
      parameters:
        - name: assignCode
          in: path
          required: true
          description: The assignment code
          schema:
            type: string
      responses:
        '200':
          description: Details for assignment successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#components/schemas/Submission'
                type: array
        '404':
          description: Submissions not found
        '500':
          description: Internal server error
...